pyproject.toml
[tool.poetry]
name = "ninja-python-django"
version = "0.1.0"
description = ""
authors = ["yuhexiong <yuhexiong871210@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
django-ninja = "^1.3.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
